PR: Fix Load Testing Framework and TokenUsageLog SQLite Bug closes #1225 (#1226)
* Load testing


* Load testing


---------

+++ mcpgateway/db.py
@@ -29,7 +29,7 @@ import uuid
 
 # Third-Party
 import jsonschema
-from sqlalchemy import BigInteger, Boolean, Column, create_engine, DateTime, event, Float, ForeignKey, func, Index, Integer, JSON, make_url, select, String, Table, Text, UniqueConstraint
+from sqlalchemy import Boolean, Column, create_engine, DateTime, event, Float, ForeignKey, func, Index, Integer, JSON, make_url, select, String, Table, Text, UniqueConstraint
 from sqlalchemy.event import listen
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.hybrid import hybrid_property
@@ -2948,7 +2948,7 @@ class TokenUsageLog(Base):
     __tablename__ = "token_usage_logs"
 
     # Primary key
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     # Token reference
     token_jti: Mapped[str] = mapped_column(String(36), nullable=False, index=True)
